Keep `Vertex` and `GlobalVertex` in same module
